requirements.txt
# Core web framework
fastapi>=0.119.0
uvicorn>=0.37.0

# Image processing
Pillow>=11.3.0

# PyTorch and TorchVision (for models and transforms)
torch>=2.8.0
torchvision>=0.23.0

# For HTTP requests
requests>=2.32.5

# Templating
Jinja2>=3.1.6

# Google Gemini API
google-generativeai>=0.8.5

# Async HTTP (for Gemini if needed)
aiohttp>=3.9.3

# Environment variable management
python-dotenv>=1.0.0
